// src/app/models/student/fee-particulars.model.ts
export class FeeParticulars {
    constructor(){}
    id?: number;
    instituteId?: number;
    name?: string;
    tuitionFee?: number;
    stationery?: number;
    sports?: number;
    annualFee?: number;
    electricity?: number;
    maintenance?: number;
    miscellaneous?: number;
    createdAt?: string;
    updatedAt?: string;
}